Refactor parser module layout: organize element-specific code
- Create elements/ subdirectory with annotations.rs and foreign.rs
- Move annotation parsing from document.rs to elements/annotations.rs
- Move foreign block parsing from combinators.rs to elements/foreign.rs
- Move build_document_content_parser from document.rs to parser.rs
- Keep document.rs focused on document-level parsing (delegates to parser.rs)
- Move element-specific tests to their respective element modules
- Update all imports and module declarations
- All tests pass
