+++ src/txxt_nano/parser/elements/mod.rs
@@ -1,0 +1,7 @@
+//! Element-specific parsers
+//!
+//! This module contains parsing logic for individual txxt elements.
+//! Each element type has its own module with its parser implementation and tests.
+
+pub mod annotations;
+pub mod foreign;
